farah_clinic-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic[email]==2.5.0
reportlab==4.0.7
python-multipart==0.0.6
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz==2023.3
email-validator==2.1.0
